cleanup: split up cdtext module misc: handle genre/code splitting at parse misc: treat unused/unspecified genre codes as no code
