+++ riprip_core/skel/cdtext/6708210a.cdtext.txt
@@ -5,7 +5,6 @@ DISC:
 	MESSAGE: IL2128
 	BARCODE: 8-10874-02762-7
 	DISC ID: 810874027627
-	GENRE CODE: 1 (Unspecified)
 TRACK 01:
 	TITLE: Funny Music
 	ISRC: USD8D2528001
